+++ src/scipp/__init__.py
@@ -117,7 +117,7 @@ setattr(Variable, 'plot', plot)
 setattr(DataArray, 'plot', plot)
 setattr(Dataset, 'plot', plot)
 
-from .core.coord_factory import _bin, _group, _hist, _nanhist
+from .core.binning import _bin, _group, _hist, _nanhist
 setattr(DataArray, 'bin', _bin)
 setattr(DataArray, 'group', _group)
 setattr(DataArray, 'hist', _hist)
