+++ src/scipp/core/binning.py
